Move source to `src` directory
