+++ src/fights/base.py
