+++ src/fights/envs/quoridor_cython.pyx
@@ -185,7 +185,7 @@ cdef int _is_moving_legal(int [:,:,:] board_view, int x, int y, int agent_id, in
 
     return 1
 
-def legal_actions(state, int agent_id, int board_size):
+def fast_legal_actions(state, int agent_id, int board_size):
     
     cdef int dir_id, action_type, next_pos_x, next_pos_y, cx, cy, nowpos_x, nowpos_y
     cdef int directions[12][2]
